==> 2_RunFS.ipynb <==
import sys
import os
import time
import pickle
import pandas as pd
import numpy as np
from datetime import datetime
from tqdm import tqdm
import matplotlib.pyplot as plt
import seaborn as sns
# Local imports
from Utils.utils import MinMaxScale_datasets
from Utils.feature_selection_methods import (
    select_n_features_LC, select_n_features_MI,
    select_n_features_DT, select_n_features_MLPSHAP
)

# -------------------------- Configuration -------------------------- #
farm = 'Kelmarsh'
target_feature = 'Generator bearing rear temperature (°C)'

# -------------------------- Load Data -------------------------- #
with open(f'./1_healthy_datasets/{farm}_HealthyDatasets.pkl', 'rb') as f:
    healthy_datasets = pickle.load(f)

# Normalize datasets (excluding target variable)
healthy_datasets = MinMaxScale_datasets(healthy_datasets, target_feature)

# Select datasets to process
dataset_names = [list(healthy_datasets.keys())[0]]
print('Dataset names:', dataset_names)

# Define feature selection methods
fs_methods = {
    'MLPSHAP': select_n_features_MLPSHAP,
    'LC': select_n_features_LC,
    'MI': select_n_features_MI,
    'DT': select_n_features_DT
}

# Output directory: where cleaned files will be saved
save_dir = './2_ordered_inputs'
os.makedirs(save_dir, exist_ok=True)

# -------------------------- Feature Selection Loop -------------------------- #
chosen_features_dict = {}
start_time = time.time()

for dataset_name in dataset_names:
    print(f'\nProcessing dataset: {dataset_name}')

    # Load split sets
    train_set = healthy_datasets[dataset_name]['train']
    valid_set = healthy_datasets[dataset_name]['valid']
    test_set = healthy_datasets[dataset_name]['test']

    total_features = train_set.shape[1]

    for fs_method, selector in fs_methods.items():
        print(f'  Feature Selection Method: {fs_method}')

        # Copy data
        selector_set = train_set.copy()
        valid_copy = valid_set.copy()

        # Remove rows with NaNs from selector set
        nan_rows = selector_set.isna().any(axis=1)
        selector_set = selector_set.loc[~nan_rows]

        if fs_method == 'MLPSHAP':
            # Perform feature selection (using all features for full ranking)
            chosen_features, weights = selector(selector_set, valid_copy, target_feature, total_features, epochs=200, background_samples = 400)
        else:
            chosen_features, weights = selector(selector_set, target_feature, total_features)          

        # Store selected features
        key = f'{dataset_name}_{fs_method}'
        chosen_features_dict[key] = list(chosen_features)

    # -------------------------- Save Results -------------------------- #
    with open(f'{save_dir}/{farm}_ordered_inputs.pkl', 'wb') as f:
        pickle.dump(chosen_features_dict, f)
    
    print(f'\nFeature selection completed and saved to: {output_path}')
